+++ app/src/main/java/com/shometeam/ao/shome/Screens/MainPart/MainRecyclerAdapter.java
@@ -41,27 +41,27 @@ public class MainRecyclerAdapter  extends RecyclerView.Adapter<MainRecyclerAdapt
 
     @Override
     public void onBindViewHolder(MainRecyclerAdapter.ViewHolder holder, int position) {
-        ArrayList<Double> values = new ArrayList<>();
-        values.add(33d);
-        values.add(37d);
-        values.add(32d);
-        values.add(31d);
-        values.add(39d);
-        values.add(22d);
-        values.add(44d);
-        values.add(35d);
-        values.add(21d);
+        ArrayList<Float> values = new ArrayList<>();
+        values.add(33f);
+        values.add(37f);
+        values.add(32f);
+        values.add(31f);
+        values.add(39f);
+        values.add(22f);
+        values.add(44f);
+        values.add(35f);
+        values.add(21f);
 
-        ArrayList<Double> cValues = new ArrayList<>();
-        cValues.add(35d);
-        cValues.add(39d);
-        cValues.add(35d);
-        cValues.add(39d);
-        cValues.add(44d);
-        cValues.add(25d);
-        cValues.add(47d);
-        cValues.add(37d);
-        cValues.add(25d);
+        ArrayList<Float> cValues = new ArrayList<>();
+        cValues.add(35f);
+        cValues.add(39f);
+        cValues.add(35f);
+        cValues.add(39f);
+        cValues.add(44f);
+        cValues.add(25f);
+        cValues.add(47f);
+        cValues.add(37f);
+        cValues.add(25f);
 
         holder.mGraphiView = holder.mGraphiView
                 .withViewType(mDataset.get(position).mType)
